Complete Epic #522 Phase D: docs, test cleanup, input registration
ORD-10:
- Rewrite ludots-selection-order-dataflow.html sections for EntityView/OrderQueue path
- Add gitbook/architecture/interaction-view-order-massnav-boundary.md
- Update tracker, RFC acceptance checklist, entity_selection_architecture, formal chain

ORD-11:
- Rename MassNavigationLocalCommandInputSystemTests -> MassNavigationMoveOrderSubmitterTests
- Extract MassNavigationTestConfigFactory; align test setup with EntityView-only selection

Architecture close-out:
- Register MassNavigationPrimarySelectionViewBootstrapSystem and MassNavigationMoveOrderSourceSystem from CoreInputMod
- Remove Input system registration from MassNavigationRuntime
- Delete dead MassNavigationInputActions and zombie SelectionViewKey service keys
