--- src/Core/MassNavigation/MassNavigationInputActions.cs
@@ -1,8 +1,0 @@
-namespace Ludots.Core.MassNavigation.Input;
-
-internal static class MassNavigationInputActions
-{
-    public const string RotateLeft = "MassNavigation_RotateLeft";
-    public const string RotateRight = "MassNavigation_RotateRight";
-    public const string ResetScene = "MassNavigation_ResetScene";
-}
